Create syntaxer
Confirms that the syntax of the program is correct, and can be correctly transformed into tokens in the lexer. Checks open braces/brackets, and confirms there is an even amount. Checks spelling to confirm every word is correct. This can be run prebuild, and will be used to print error messages.